+++ lib/services/auth_service.dart
@@ -1,12 +1,12 @@
 import 'package:dio/dio.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:get_storage/get_storage.dart';
-
 import 'package:mi_utem/models/carrera.dart';
 import 'package:mi_utem/models/usuario.dart';
 import 'package:mi_utem/services/carreras_service.dart';
 import 'package:mi_utem/services/perfil_service.dart';
 import 'package:mi_utem/utils/dio_miutem_client.dart';
+import 'package:sentry_flutter/sentry_flutter.dart';
 
 class AuthService {
   static final Dio _dio = DioMiUtemClient.initDio;
@@ -95,6 +95,9 @@ class AuthService {
           version == versionCorrecta;
 
       if (!isLoggedIn) {
+        Sentry.configureScope(
+          (scope) => scope.setUser(null),
+        );
         await logOut();
       }