+++ lib/widgets/custom_app_bar.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-
 import 'package:mi_utem/themes/theme.dart';
 
 class CustomAppBar extends AppBar {
@@ -238,7 +237,7 @@ class CustomAppBar extends AppBar {
   /// The typographic styles to use for text in the app bar. Typically this is
   /// set along with [brightness] [backgroundColor], [iconTheme].
   ///
-  /// If this property is null, then [AppBarTheme.textTheme] of
+  /// If this property is null, then [AppBarTheme] of
   /// [ThemeData.appBarTheme] is used. If that is also null, then
   /// [ThemeData.primaryTextTheme] is used.
   final TextTheme? textTheme;